+++ packages/components/fields/localized-multiline-text-field/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './localized-multiline-text-field';
+
+export { default as version } from './version';
