+++ packages/components/fields/localized-multiline-text-field/src/localized-multiline-text-field.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent } from '../../../../test/test-utils';
+import { render, fireEvent } from '../../../../../test/test-utils';
 import LocalizedMultilineTextField from './localized-multiline-text-field';
 
 // This component is used to enable easy testing.
