+++ packages/components/fields/localized-multiline-text-field/src/localized-multiline-text-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { LocalizedMultilineTextField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const lorem =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.';
